mv helper functions to the utils
