+++ crates/sdk-derive/src/solidity_router.rs
@@ -1,26 +1,27 @@
-use crate::utils::{get_all_methods, get_public_methods};
-use convert_case::{Case, Casing};
+use crate::utils::{
+    get_all_methods,
+    get_public_methods,
+    parse_function_inputs,
+    rust_name_to_sol,
+    rust_type_to_sol,
+    sol_call_fn_name,
+};
+use convert_case::Casing;
 use proc_macro::TokenStream;
 use quote::{quote, ToTokens};
 use syn::{
     self,
     parse::Parse,
     parse_macro_input,
-    punctuated::Punctuated,
-    Expr,
     FnArg,
     Ident,
     ImplItem,
     ImplItemFn,
     ItemImpl,
-    Lit,
     LitStr,
-    Token,
-    Type,
-    TypePath,
 };
 
-// #[proc_macro_attribute]
+// #[proc_macrot_attribute]
 pub fn derive_solidity_router(_attr: TokenStream, item: TokenStream) -> TokenStream {
     let ast: ItemImpl = parse_macro_input!(item as ItemImpl);
     let struct_name = &ast.self_ty;
@@ -226,130 +227,10 @@ fn derive_route_selector_args(
     }
 }
 
-fn parse_function_inputs(inputs: &Punctuated<FnArg, Token![,]>) -> Vec<proc_macro2::TokenStream> {
-    inputs
-        .iter()
-        .filter_map(|arg| {
-            if let FnArg::Typed(pat_type) = arg {
-                if let syn::Pat::Ident(pat_ident) = &*pat_type.pat {
-                    let param_name = &pat_ident.ident;
-                    let sol_type = rust_type_to_sol(&*pat_type.ty);
-                    Some(quote! { #sol_type #param_name })
-                } else {
-                    None
-                }
-            } else {
-                None
-            }
-        })
-        .collect()
-}
-
-fn sol_call_fn_name(method_name: &Ident) -> Ident {
-    let method_name_sol = rust_name_to_sol(method_name);
-    Ident::new(&(method_name_sol.to_string() + "Call"), method_name.span())
-}
-
-fn rust_name_to_sol(ident: &Ident) -> Ident {
-    let span = ident.span();
-    let camel_name = ident.to_string().to_case(Case::Camel);
-    Ident::new(&camel_name, span)
-}
-
-fn rust_type_to_sol(ty: &Type) -> proc_macro2::TokenStream {
-    match ty {
-        Type::Array(ty) => convert_array_type(ty),
-        Type::Paren(ty) => convert_paren_type(ty),
-        Type::Slice(ty) => convert_slice_type(ty),
-        Type::Tuple(ty) => convert_tuple_type(ty),
-        Type::Path(type_path) => convert_path_type(type_path),
-        Type::Reference(type_ref) => rust_type_to_sol(&type_ref.elem),
-        _ => panic!("Unsupported type: {}", ty.to_token_stream().to_string()),
-    }
-}
-
-fn convert_array_type(ty: &syn::TypeArray) -> proc_macro2::TokenStream {
-    if let Expr::Lit(expr_lit) = &ty.len {
-        if let Lit::Int(lit_int) = &expr_lit.lit {
-            let len = lit_int.base10_digits();
-            let len_token: proc_macro2::TokenStream = len.parse().expect("Invalid token");
-            let elem_type = rust_type_to_sol(&ty.elem);
-            quote! { #elem_type[#len_token] }
-        } else {
-            panic!("Unsupported array length literal")
-        }
-    } else {
-        panic!("Unsupported array length expression");
-    }
-}
-
-fn convert_paren_type(ty: &syn::TypeParen) -> proc_macro2::TokenStream {
-    rust_type_to_sol(&ty.elem)
-}
-
-fn convert_slice_type(ty: &syn::TypeSlice) -> proc_macro2::TokenStream {
-    let elem_type = rust_type_to_sol(&ty.elem);
-    quote! { #elem_type[] }
-}
-
-fn convert_tuple_type(ty: &syn::TypeTuple) -> proc_macro2::TokenStream {
-    let elems = ty.elems.iter().map(|elem| rust_type_to_sol(elem));
-    quote! { (#(#elems),*)}
-}
-
-fn convert_path_type(type_path: &TypePath) -> proc_macro2::TokenStream {
-    let ident = &type_path.path.segments.last().unwrap().ident;
-    match ident.to_string().as_str() {
-        "String" | "str" => quote! { string },
-        "bool" => quote! { bool },
-        "u8" => quote! { uint8 },
-        "u16" => quote! { uint16 },
-        "u32" => quote! { uint32 },
-        "u64" => quote! { uint64 },
-        "u128" => quote! { uint128 },
-        "u256" | "uint" => quote! { uint256 },
-        "U256" => quote! { uint256 },
-        "i8" => quote! { int8 },
-        "i16" => quote! { int16 },
-        "i32" => quote! { int32 },
-        "i64" => quote! { int64 },
-        "i128" => quote! { int128 },
-        "i256" | "int" => quote! { int256 },
-        "Address" => quote! { address },
-        "Bytes" => quote! { bytes },
-        "Vec" => {
-            if let syn::PathArguments::AngleBracketed(args) =
-                &type_path.path.segments.last().unwrap().arguments
-            {
-                if let Some(syn::GenericArgument::Type(arg_ty)) = args.args.first() {
-                    let elem_type = rust_type_to_sol(arg_ty);
-                    quote! { #elem_type[] }
-                } else {
-                    panic!("Unsupported vector element type")
-                }
-            } else {
-                panic!("Unsupported vector arguments")
-            }
-        }
-        ident_str if ident_str.starts_with("bytes") => {
-            if ident_str == "bytes" {
-                quote! { bytes }
-            } else if ident_str.len() > 5 && ident_str[5..].parse::<usize>().is_ok() {
-                let bytes_len: usize = ident_str[5..].parse().unwrap();
-                let bytes_ident = Ident::new(&format!("bytes{}", bytes_len), ident.span());
-                quote! { #bytes_ident }
-            } else {
-                panic!("Unsupported bytes type: {}", ident_str)
-            }
-        }
-        _ => panic!("Unsupported type: {}", ident),
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use syn::{parse_quote, Ident, TypeArray, TypeParen, TypePath, TypeSlice, TypeTuple};
+    use syn::{parse_quote, Ident};
 
     #[test]
     fn test_get_signatures_full_signature() {
@@ -488,78 +369,5 @@ mod tests {
         let method_name = Ident::new("test_function", proc_macro2::Span::call_site());
         let method_call_ident = sol_call_fn_name(&method_name);
         assert_eq!(method_call_ident.to_string(), "testFunctionCall");
-    }
-
-    #[test]
-    fn test_convert_array_type() {
-        let ty: TypeArray = parse_quote!([u8; 32]);
-        let result = convert_array_type(&ty);
-        assert_eq!(result.to_string(), "uint8 [32]");
-    }
-
-    #[test]
-    fn test_convert_paren_type() {
-        let ty: TypeParen = parse_quote!((u8));
-        let result = convert_paren_type(&ty);
-        assert_eq!(result.to_string(), "uint8");
-    }
-
-    #[test]
-    fn test_convert_slice_type() {
-        let ty: TypeSlice = parse_quote!([u8]);
-        let result = convert_slice_type(&ty);
-        assert_eq!(result.to_string(), "uint8 []");
-    }
-
-    #[test]
-    fn test_convert_tuple_type() {
-        let ty: TypeTuple = parse_quote!((u8, u16));
-        let result = convert_tuple_type(&ty);
-        assert_eq!(result.to_string(), "(uint8 , uint16)");
-    }
-
-    #[test]
-    fn test_convert_path_type_string() {
-        let ty: TypePath = parse_quote!(String);
-        let result = convert_path_type(&ty);
-        assert_eq!(result.to_string(), "string");
-        let ty: TypePath = parse_quote!(str);
-        let result = convert_path_type(&ty);
-        assert_eq!(result.to_string(), "string");
-    }
-
-    #[test]
-    fn test_convert_path_type_bool() {
-        let ty: TypePath = parse_quote!(bool);
-        let result = convert_path_type(&ty);
-        assert_eq!(result.to_string(), "bool");
-    }
-
-    #[test]
-    fn test_convert_path_type_uint() {
-        let ty: TypePath = parse_quote!(u256);
-        let result = convert_path_type(&ty);
-        assert_eq!(result.to_string(), "uint256");
-    }
-
-    #[test]
-    fn test_convert_path_type_vec() {
-        let ty: TypePath = parse_quote!(Vec<u8>);
-        let result = convert_path_type(&ty);
-        assert_eq!(result.to_string(), "uint8 []");
-    }
-
-    #[test]
-    fn test_convert_path_type_bytes() {
-        let ty: TypePath = parse_quote!(bytes);
-        let result = convert_path_type(&ty);
-        assert_eq!(result.to_string(), "bytes");
-    }
-
-    #[test]
-    fn test_convert_path_type_bytes_fixed() {
-        let ty: TypePath = parse_quote!(bytes32);
-        let result = convert_path_type(&ty);
-        assert_eq!(result.to_string(), "bytes32");
     }
 }
